🎨📣 Affiche le logo bien en vu dans les intégrations iframes
