+++ site/source/components/SchemeComparaison.tsx
@@ -19,7 +19,8 @@ import styled, { css } from 'styled-components'
 import dirigeantComparaison from '../pages/Simulateurs/configs/rémunération-dirigeant.yaml'
 import SeeAnswersButton from './conversation/SeeAnswersButton'
 import PeriodSwitch from './PeriodSwitch'
-import { SimulationGoal, SimulationGoals } from './SimulationGoals'
+import { SimulationGoals, SimulationGoal } from './Simulation'
+
 import Emoji from './utils/Emoji'
 import { useEngine } from './utils/EngineContext'
 import useSimulationConfig from './utils/useSimulationConfig'
