+++ site/source/iframe-integration-script.js
@@ -1,6 +1,5 @@
 import { iframeResizer } from 'iframe-resizer'
-import logoFrSvg from 'Images/logo-monentreprise.svg'
-import logoEnSvg from 'Images/logo-mycompany.svg'
+
 import { hexToHSL } from './hexToHSL'
 
 let script =
@@ -63,26 +62,13 @@ params.append('utm_medium', 'iframe')
 params.append('utm_campaign', 'newtext')
 url.search = params.toString()
 const simulateurURL = url.toString()
-url.pathname = '/simulateurs'
-const simulateursURL = url.toString()
-url.pathname = '/'
-const monEntrepriseUrl = url.toString()
+
 links.innerHTML = `
 	<div style="text-align: center; margin-bottom: 2rem; font-size: 80%">
-	Ce simulateur a été créé par
+	Retrouvez ce simulateur et bien d'autres sur 
 	<a href="${simulateurURL}">
 		mon-entreprise.urssaf.fr
 	</a><br/>
-	Découvrez l'ensemble des simulateurs disponibles <a href="${simulateursURL}">ici</a><br/>
-
-	<a href="${monEntrepriseUrl}">
-			<img
-				style="height: 40px; margin: 10px"
-				src="${process.env.FR_BASE_URL + '/' + (lang === 'fr' ? logoFrSvg : logoEnSvg)}"
-				alt="mon-entreprise.urssaf.fr : l'assistant officiel du créateur d'entreprise"
-			/>
-		</a>
-
 	</div>
 `
