+++ site/source/pages/Simulateurs/ArtisteAuteur.tsx
@@ -4,7 +4,7 @@ import PeriodSwitch from 'Components/PeriodSwitch'
 import SimulateurWarning from 'Components/SimulateurWarning'
 import Simulation from 'Components/Simulation'
 import { InstitutionsPartenairesArtisteAuteur } from 'Components/simulationExplanation/InstitutionsPartenaires'
-import { SimulationGoal, SimulationGoals } from 'Components/SimulationGoals'
+import { SimulationGoal, SimulationGoals } from 'Components/Simulation'
 import { EngineContext } from 'Components/utils/EngineContext'
 import useSimulationConfig from 'Components/utils/useSimulationConfig'
 import { H2 } from 'DesignSystem/typography/heading'
