+++ site/source/pages/Simulateurs/AutoEntrepreneur.tsx
@@ -4,7 +4,7 @@ import PeriodSwitch from 'Components/PeriodSwitch'
 import SimulateurWarning from 'Components/SimulateurWarning'
 import Simulation from 'Components/Simulation'
 import { InstitutionsPartenairesAutoEntrepreneur } from 'Components/simulationExplanation/InstitutionsPartenaires'
-import { SimulationGoal, SimulationGoals } from 'Components/SimulationGoals'
+import { SimulationGoal, SimulationGoals } from 'Components/Simulation'
 import StackedBarChart from 'Components/StackedBarChart'
 import { H2 } from 'DesignSystem/typography/heading'
 import { useContext } from 'react'
