+++ site/source/pages/Simulateurs/Dividendes.tsx
@@ -3,7 +3,7 @@ import { HiddenOptionContext } from 'Components/conversation/ChoicesInput'
 import { Condition } from 'Components/EngineValue'
 import Notifications from 'Components/Notifications'
 import Simulation from 'Components/Simulation'
-import { SimulationGoal, SimulationGoals } from 'Components/SimulationGoals'
+import { SimulationGoal, SimulationGoals } from 'Components/Simulation'
 import StackedBarChart from 'Components/StackedBarChart'
 import Warning from 'Components/ui/WarningBlock'
 import { useEngine } from 'Components/utils/EngineContext'
