+++ site/source/pages/Simulateurs/ImpôtSociété.tsx
@@ -2,7 +2,7 @@ import { updateSituation } from 'Actions/actions'
 import RuleInput from 'Components/conversation/RuleInput'
 import Value from 'Components/EngineValue'
 import Notifications from 'Components/Notifications'
-import { SimulationGoal, SimulationGoals } from 'Components/SimulationGoals'
+import { SimulationGoal, SimulationGoals } from 'Components/Simulation'
 import { FromTop } from 'Components/ui/animate'
 import Warning from 'Components/ui/WarningBlock'
 import useSimulationConfig from 'Components/utils/useSimulationConfig'
