+++ site/source/pages/Simulateurs/Salarié.tsx
@@ -4,7 +4,7 @@ import PeriodSwitch from 'Components/PeriodSwitch'
 import RuleLink from 'Components/RuleLink'
 import Simulation from 'Components/Simulation'
 import SalaryExplanation from 'Components/simulationExplanation/SalaryExplanation'
-import { SimulationGoal, SimulationGoals } from 'Components/SimulationGoals'
+import { SimulationGoal, SimulationGoals } from 'Components/Simulation'
 import { FromTop } from 'Components/ui/animate'
 import Emoji from 'Components/utils/Emoji'
 import { useEngine } from 'Components/utils/EngineContext'
